Polkadot v0.9.37 (#299)
* update dependencies

* fix isssues 1

* fix weights

* fix issues 2

* fix tests build

* fmt

* fix clippy 1

* fix clippy 2

* fix clippy 3

* fmt

* install protoc in github workflows

+++ ceremonies/meetup-validation/src/meetup_scenario_tests.rs
@@ -344,7 +344,7 @@ fn n_honest_participants_dont_have_internet_and_n_attackers_try_to_provoke_early
 	let mut participant_votes = vec![num_honest_participants as u32; meetup_size - n];
 
 	// n participants do not have internet
-	participant_votes.append(&mut vec![0 as u32; n]);
+	participant_votes.append(&mut vec![0_u32; n]);
 
 	let mut participant_attestations: Attestations = vec![];
